Update README.md
I am looking to manupulate UI for uplaoding an image to PDF.
Currently I have a button set up with the JavaScript below, which allows the user to click on the button to upload an image. However, the files on the users drive always appear ideal and they can not reupload after choosing an image. I want jpg and png permitence. any and all help is appreciated.